feat: Setup the Chart Object as MVP with Hatching
The key insight was to use the Country-Category pairs
as the stacks and to give the value as null for all the
data array values except the one corresponding to the current
stack. There are still some artifacts to work out here---lots
of zero bars at the bottom of the chart, and weird spacing because of
it, but it's not so bad at the moment.